requirements.txt
setuptools
python-dotenv
requests
numpy
pandas
matplotlib
scikit-learn
seaborn
jupyter-dash
jupyterlab
openai
ollama
google-generativeai
anthropic
gradio
transformers
datasets
diffusers
torch
langchain
langchain-community
langchain_openai
langchain_chroma